Change battery percentage floats to doubles

+++ src/Ipr.WaterSensor/Ipr.WaterSensor.Core/Entities/WaterLevel.cs
@@ -9,7 +9,7 @@ namespace Ipr.WaterSensor.Core.Entities
 {
     public class WaterLevel : BaseEntity
     {
-        public float Percentage { get; set; }
+        public double Percentage { get; set; }
         public DateTime DateTimeMeasured { get; set; }
     }
 }
